Add d3 dependency. Moved auxiliar ts
